change database

+++ C#/DaisyDBProject/DaisyDBProject/Models/LeaveMessage.cs
@@ -9,7 +9,7 @@ namespace DaisyDBProject.Models
         public string Account { get; set; }
         public string Time { get; set; }
         public string Content { get; set; }
-        public decimal? ReadTag { get; set; }
+        public int? ReadTag { get; set; }
 
         public virtual Users AccountNavigation { get; set; }
     }
