+++ C#/DaisyDBProject/DaisyDBProject/Models/UserNotifi.cs
@@ -8,7 +8,7 @@ namespace DaisyDBProject.Models
         public string ProjectId { get; set; }
         public string NotifiId { get; set; }
         public string Account { get; set; }
-        public decimal? ReadTag { get; set; }
+        public int? ReadTag { get; set; }
 
         public virtual Users AccountNavigation { get; set; }
         public virtual Notification Notification { get; set; }
